+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.68 (not yet released)
 ------------
 
+* Changed actions Send Email and Send pdf by email in such a way that one can set a custom from name and e-mail address. And also an alternative recipient.
+
 Version 1.67
 ------------